+++ libMachO/MachO_ABI/LoadCommands/_load_command_linkedit.c
@@ -31,13 +31,13 @@
 size_t
 _mk_load_command_type_linkedit_description(mk_load_command_ref load_command, char *output, size_t output_len)
 {
-    return snprintf(output, output_len, "<%s %p> {\n\
+    return (size_t)snprintf(output, output_len, "<%s %p> {\n\
 \tdataoff = 0x%" PRIi32 "\n\
 \tdatasize = 0x%" PRIi32 "\n\
 }",
-                    mk_type_name(load_command.type), load_command.type,
-                    _mk_load_command_type_linkedit_get_dataoff(load_command),
-                    _mk_load_command_type_linkedit_get_datasize(load_command));
+                            mk_type_name(load_command.type), load_command.type,
+                            _mk_load_command_type_linkedit_get_dataoff(load_command),
+                            _mk_load_command_type_linkedit_get_datasize(load_command));
 }
 
 //|++++++++++++++++++++++++++++++++++++|//
