+++ libMachO/MachO_ABI/LoadCommands/load_command_prebind_cksum.c
@@ -31,11 +31,11 @@
 static size_t
 _mk_load_command_prebind_cksum_copy_description(mk_load_command_ref load_command, char *output, size_t output_len)
 {
-    return snprintf(output, output_len, "<%s %p> {\n\
+    return (size_t)snprintf(output, output_len, "<%s %p> {\n\
 \tcksum: 0x%" PRIx32 "\n\
 }",
-                    mk_type_name(load_command.type), load_command.type,
-                    mk_load_command_prebind_cksum_get_cksum(load_command));
+                            mk_type_name(load_command.type), load_command.type,
+                            mk_load_command_prebind_cksum_get_cksum(load_command));
 }
 
 const struct _mk_load_command_vtable _mk_load_command_prebind_cksum_class = {
